`order::Order` no longer holds references
When mutable references to repos were involved, these circular references caused an exclusive borrow for that reference. The end result is that a repo could not be used if a mutable operation is performed on it.

Additionally, also changed the repo  to generally prefer namespaced names for order, menu, and layout, instead of using them.