Set up skeleton HTTP endpoint
A substantial amount of time was spent looking at warp. As an experiment, I wanted to hook up versioning via an HTTP header in the simplest way possible, and I found it too cumbersome to do it in warp.

A substantial amount of time was also spent finding a clean way to do this versioning in axum. It was still relatively obnoxious to find a clean implementation, but relatively clean it is!