Refactor everything to have a typed id
I kinda knew I'd do this in the end, but I wanted to see how it went. It was pretty silly for the web api to first have to query the id for some item, then pass it along. It was more operations than needed.

Instead, there are now `layout::TableId`, `menu::Id`, and `order::Id`. These can be passed into appropriate methods just fine, and, if consumer happens to have the full object, it's just a simple call to `.id()`.

Additionally, also moved the in-memory database to the `restaurant` crate and updated tests to use it.